requirements.txt
pyenchant==2.0.0
